# exports/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/events": {
      "get": {
        "operationId": "EventController_getAllEvents",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by event category",
            "schema": {
              "type": "string",
              "enum": [
                "Concerts",
                "Theater",
                "Sports",
                "Culture",
                "Comedy",
                "Festival"
              ]
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "description": "Filter by event format",
            "schema": {
              "type": "string",
              "enum": [
                "physical",
                "online",
                "hybrid"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status. Leave as default for public browsing.",
            "schema": {
              "default": "published",
              "type": "string",
              "enum": [
                "draft",
                "published",
                "cancelled",
                "completed"
              ]
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "description": "Filter by venue city. Case-insensitive partial match.",
            "schema": {
              "example": "Lagos",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Filter by venue state. Case-insensitive partial match.",
            "schema": {
              "example": "Lagos",
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum ticket price filter",
            "schema": {
              "minimum": 0,
              "example": 5000,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum ticket price filter",
            "schema": {
              "minimum": 0,
              "example": 50000,
              "type": "number"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Return events starting on or after this date",
            "schema": {
              "example": "2026-01-01T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Return events starting on or before this date",
            "schema": {
              "example": "2026-12-31T00:00:00Z",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Full-text search across title and description",
            "schema": {
              "example": "Burna Boy",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Pagination page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "default": "date",
              "type": "string",
              "enum": [
                "date",
                "price",
                "popular",
                "newest"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Sort direction",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "isFeatured",
            "required": false,
            "in": "query",
            "description": "Return only featured events",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isLive",
            "required": false,
            "in": "query",
            "description": "Return only currently live events",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "excludeSoldOut",
            "required": false,
            "in": "query",
            "description": "Exclude fully sold out events",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of published events"
          }
        },
        "summary": "Get all published events",
        "tags": [
          "Events"
        ]
      },
      "post": {
        "operationId": "EventController_createEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — creator role required"
          },
          "409": {
            "description": "Slug already exists"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create a new event (starts as draft)",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/featured": {
      "get": {
        "operationId": "EventController_getFeaturedEvents",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max results to return",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of featured events"
          }
        },
        "summary": "Get featured events",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/upcoming": {
      "get": {
        "operationId": "EventController_getUpcomingEvents",
        "parameters": [
          {
            "name": "city",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of upcoming events"
          }
        },
        "summary": "Get upcoming events",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/user/my-events": {
      "get": {
        "operationId": "EventController_getUserEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User ticket events"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get events the current user has tickets for",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/images": {
      "post": {
        "operationId": "EventController_uploadEventImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG, PNG or WebP. Max 5MB."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image uploaded, URL appended to event"
          },
          "400": {
            "description": "Invalid file type, size exceeded, or 10 image limit reached"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Upload an image to an event",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/creator/my": {
      "get": {
        "operationId": "EventController_getCreatorEvents",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "draft",
                "published",
                "cancelled",
                "completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Creator events list"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — creator role required"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get all events created by the authenticated creator",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/slug/{slug}": {
      "get": {
        "operationId": "EventController_getEventBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "URL-friendly event identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "summary": "Get event by slug",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/tiers/{tierId}/availability": {
      "get": {
        "operationId": "EventController_checkTicketAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tierId",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the ticket tier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quantity",
            "required": false,
            "in": "query",
            "description": "Number of tickets to check. Default: 1",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability result"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "summary": "Check ticket availability for a tier",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/tiers": {
      "post": {
        "operationId": "EventController_addTicketTier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTicketTierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tier added successfully"
          },
          "400": {
            "description": "Would exceed capacity or invalid dates"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Add a ticket tier to an event",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/tiers/{tierId}": {
      "patch": {
        "operationId": "EventController_updateTicketTier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tierId",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the ticket tier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketTierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tier updated successfully"
          },
          "400": {
            "description": "Quantity below sold count"
          },
          "404": {
            "description": "Event or tier not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update a ticket tier",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventController_removeTicketTier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tierId",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the ticket tier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tier removed successfully"
          },
          "400": {
            "description": "Cannot remove — tickets already sold"
          },
          "404": {
            "description": "Event or tier not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Remove a ticket tier",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/publish": {
      "post": {
        "operationId": "EventController_publishEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event published successfully"
          },
          "400": {
            "description": "Event already published or not ready"
          },
          "403": {
            "description": "Forbidden — not the event creator"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Publish a draft event",
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}": {
      "patch": {
        "operationId": "EventController_updateEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully"
          },
          "400": {
            "description": "Validation failed or business rule violated"
          },
          "403": {
            "description": "Forbidden — not the event creator"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update an event",
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventController_deleteEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted or cancelled"
          },
          "403": {
            "description": "Forbidden — not the event creator"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Delete event or cancel if tickets have been sold",
        "tags": [
          "Events"
        ]
      },
      "get": {
        "operationId": "EventController_getEventById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId of the event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found"
          },
          "404": {
            "description": "Event not found"
          }
        },
        "summary": "Get event by ID",
        "tags": [
          "Events"
        ]
      }
    },
...
  },
  "info": {
    "title": "Events API",
    "description": "API for creating and managing events",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddTicketTierDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tier display name",
            "example": "VIP",
            "minLength": 2,
            "maxLength": 50
          },
          "price": {
            "type": "number",
            "description": "Price per ticket. Use 0 for free tickets.",
            "example": 45000,
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "description": "ISO currency code",
            "example": "NGN",
            "default": "NGN"
          },
          "quantity": {
            "type": "number",
            "description": "Total tickets available for this tier",
            "example": 100,
            "minimum": 1
          },
          "benefits": {
            "description": "Perks included with this tier",
            "example": [
              "Access to VIP lounge",
              "Free drinks"
            ],
            "default": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "salesStart": {
            "type": "string",
            "description": "When ticket sales open for this tier",
            "example": "2026-07-01T10:00:00Z"
          },
          "salesEnd": {
            "type": "string",
            "description": "When ticket sales close. Must be strictly before event startDate.",
            "example": "2026-08-14T23:59:00Z"
          }
        },
        "required": [
          "name",
          "price",
          "quantity",
          "salesStart",
          "salesEnd"
        ]
      },
...
    }
  }
}